+++ projects/rollup-fi/index.js
@@ -1,28 +1,8 @@
 
 const { gmxExports } = require('../helper/gmx')
-const ADDRESSES = require('../helper/coreAssets.json')
-const sdk = require('@defillama/sdk')
-const { sumTokensExport } = require('../helper/unwrapLPs')
-const { getLogs } = require('../helper/cache/getLogs')
 
 module.exports = {
   era: {
-    tvl: sdk.util.sumChainTvls([
-      gmxExports({ vault: '0xefdE2AeFE307A7362C7E0E3BE019D1491Dc7E163' }),
-      sumTokensExport({ owner: '0x4992eb45172868f0d8cceb91190e159bdf571461', tokens: [ADDRESSES.era.USDC] }),
-      cryptoPoolTvl,
-    ])
+    tvl: gmxExports({ vault: '0xefdE2AeFE307A7362C7E0E3BE019D1491Dc7E163' })
   },
-};
-
-async function cryptoPoolTvl(_, _b, _cb, { api, }) {
-  const logs = await getLogs({
-    api,
-    target: '0x7C5667677e7E6d5a7a3b7cb9EF25a2B4ad2C745A',
-    eventAbi: "event NewVault(address indexed vault, address indexed creator, address token)",
-    onlyArgs: true,
-    fromBlock: 18448266,
-  })
-
-  return api.sumTokens({ tokensAndOwners: logs.map(log => [log.token, log.vault]) })
 }
